Adding new version of Aur with safety fixes
