+++ Package/pubspec.yaml
@@ -1,6 +1,6 @@
 name: aureus
 description: An open-source design system for non-profits to quickly develop user safety critical applications.
-version: 0.0.20
+version: 0.0.21
 homepage: https://withaureus.org/
 documentation: https://github.com/Astra-Labs/Aureus/
